+++ x/crosschain/keeper/msg_server_vote_outbound_tx.go
@@ -2,20 +2,13 @@ package keeper
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"math/big"
 
 	cosmoserrors "cosmossdk.io/errors"
-	"cosmossdk.io/math"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/rs/zerolog/log"
-	"github.com/zeta-chain/zetacore/common"
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
 	observerkeeper "github.com/zeta-chain/zetacore/x/observer/keeper"
-	observertypes "github.com/zeta-chain/zetacore/x/observer/types"
 )
 
 // VoteOnObservedOutboundTx casts a vote on an outbound transaction observed on a connected chain (after
@@ -99,153 +92,18 @@ func (k msgServer) VoteOnObservedOutboundTx(goCtx context.Context, msg *types.Ms
 	}
 
 	// if ballot successful, the value received should be the out tx amount
-	if ballot.BallotStatus != observertypes.BallotStatus_BallotFinalized_FailureObservation {
-		if !msg.ValueReceived.Equal(cctx.GetCurrentOutTxParam().Amount) {
-			log.Error().Msgf("VoteOnObservedOutboundTx: Mint mismatch: %s value received vs %s cctx amount",
-				msg.ValueReceived,
-				cctx.GetCurrentOutTxParam().Amount)
-			return nil, cosmoserrors.Wrap(sdkerrors.ErrInvalidRequest, fmt.Sprintf("ValueReceived %s does not match sent value %s", msg.ValueReceived, cctx.GetCurrentOutTxParam().Amount))
-		}
-	}
-
-	// Update CCTX values
-	cctx.GetCurrentOutTxParam().OutboundTxHash = msg.ObservedOutTxHash
-	cctx.GetCurrentOutTxParam().OutboundTxGasUsed = msg.ObservedOutTxGasUsed
-	cctx.GetCurrentOutTxParam().OutboundTxEffectiveGasPrice = msg.ObservedOutTxEffectiveGasPrice
-	cctx.GetCurrentOutTxParam().OutboundTxEffectiveGasLimit = msg.ObservedOutTxEffectiveGasLimit
-	cctx.CctxStatus.LastUpdateTimestamp = ctx.BlockHeader().Time.Unix()
-
-	// Fund the gas stability pool with the remaining funds
-	if err := k.FundGasStabilityPoolFromRemainingFees(ctx, *cctx.GetCurrentOutTxParam(), msg.OutTxChain); err != nil {
-		log.Error().Msgf(
-			"VoteOnObservedOutboundTx: CCTX: %s Can't fund the gas stability pool with remaining fees %s", cctx.Index, err.Error(),
-		)
-	}
-
-	tss, found := k.zetaObserverKeeper.GetTSS(ctx)
-	if !found {
-		return nil, types.ErrCannotFindTSSKeys
-	}
-
-	tmpCtx, commit := ctx.CacheContext()
-	err = func() error { //err = FinalizeOutbound(k, ctx, &cctx, msg, ballot.BallotStatus)
-		cctx.GetCurrentOutTxParam().OutboundTxObservedExternalHeight = msg.ObservedOutTxBlockHeight
-		oldStatus := cctx.CctxStatus.Status
-		switch ballot.BallotStatus {
-		case observertypes.BallotStatus_BallotFinalized_SuccessObservation:
-			switch oldStatus {
-			case types.CctxStatus_PendingRevert:
-				cctx.CctxStatus.ChangeStatus(types.CctxStatus_Reverted, "")
-			case types.CctxStatus_PendingOutbound:
-				cctx.CctxStatus.ChangeStatus(types.CctxStatus_OutboundMined, "")
-			}
-			newStatus := cctx.CctxStatus.Status.String()
-			EmitOutboundSuccess(tmpCtx, msg, oldStatus.String(), newStatus, cctx)
-		case observertypes.BallotStatus_BallotFinalized_FailureObservation:
-			if msg.CoinType == common.CoinType_Cmd || common.IsZetaChain(cctx.InboundTxParams.SenderChainId) {
-				// if the cctx is of coin type cmd or the sender chain is zeta chain, then we do not revert, the cctx is aborted
-				cctx.CctxStatus.ChangeStatus(types.CctxStatus_Aborted, "")
-			} else {
-				switch oldStatus {
-				case types.CctxStatus_PendingOutbound:
-
-					gasLimit, err := k.GetRevertGasLimit(ctx, &cctx)
-					if err != nil {
-						return errors.New("can't get revert tx gas limit" + err.Error())
-					}
-					if gasLimit == 0 {
-						// use same gas limit of outbound as a fallback -- should not happen
-						gasLimit = cctx.OutboundTxParams[0].OutboundTxGasLimit
-					}
-
-					// create new OutboundTxParams for the revert
-					revertTxParams := &types.OutboundTxParams{
-						Receiver:           cctx.InboundTxParams.Sender,
-						ReceiverChainId:    cctx.InboundTxParams.SenderChainId,
-						Amount:             cctx.InboundTxParams.Amount,
-						OutboundTxGasLimit: gasLimit,
-					}
-					cctx.OutboundTxParams = append(cctx.OutboundTxParams, revertTxParams)
-
-					err = k.PayGasAndUpdateCctx(
-						tmpCtx,
-						cctx.InboundTxParams.SenderChainId,
-						&cctx,
-						cctx.OutboundTxParams[0].Amount,
-						false,
-					)
-					if err != nil {
-						return err
-					}
-					err = k.UpdateNonce(tmpCtx, cctx.InboundTxParams.SenderChainId, &cctx)
-					if err != nil {
-						return err
-					}
-					cctx.CctxStatus.ChangeStatus(types.CctxStatus_PendingRevert, "Outbound failed, start revert")
-				case types.CctxStatus_PendingRevert:
-					cctx.CctxStatus.ChangeStatus(types.CctxStatus_Aborted, "Outbound failed: revert failed; abort TX")
-				}
-			}
-			newStatus := cctx.CctxStatus.Status.String()
-			EmitOutboundFailure(ctx, msg, oldStatus.String(), newStatus, cctx)
-		}
-		return nil
-	}()
+	err = k.GetOutbound(ctx, &cctx, *msg, ballot.BallotStatus)
 	if err != nil {
-		// do not commit tmpCtx
-		cctx.CctxStatus.ChangeStatus(types.CctxStatus_Aborted, err.Error())
-		cctx.GetCurrentOutTxParam().TxFinalizationStatus = types.TxFinalizationStatus_Executed
-		ctx.Logger().Error(err.Error())
-		// #nosec G701 always in range
-		k.GetObserverKeeper().RemoveFromPendingNonces(ctx, tss.TssPubkey, msg.OutTxChain, int64(msg.OutTxTssNonce))
-		k.RemoveOutTxTracker(ctx, msg.OutTxChain, msg.OutTxTssNonce)
-		k.SetCctxAndNonceToCctxAndInTxHashToCctx(ctx, cctx)
+		return nil, err
+	}
+	// Fund the gas stability pool with the remaining funds
+	k.FundStabiltityPool(ctx, &cctx)
+
+	err = k.ProcessOutbound(ctx, &cctx, ballot.BallotStatus, msg.ValueReceived.String())
+	if err != nil {
+		k.SaveFailedOutBound(ctx, &cctx, msg.ValueReceived.String())
 		return &types.MsgVoteOnObservedOutboundTxResponse{}, nil
 	}
-	commit()
-	// Set the ballot index to the finalized ballot
-	cctx.GetCurrentOutTxParam().OutboundTxBallotIndex = ballotIndex
-	cctx.GetCurrentOutTxParam().TxFinalizationStatus = types.TxFinalizationStatus_Executed
-	// #nosec G701 always in range
-	k.GetObserverKeeper().RemoveFromPendingNonces(ctx, tss.TssPubkey, msg.OutTxChain, int64(msg.OutTxTssNonce))
-	k.RemoveOutTxTracker(ctx, msg.OutTxChain, msg.OutTxTssNonce)
-	ctx.Logger().Info(fmt.Sprintf("Remove tracker %s: , Block Height : %d ", getOutTrackerIndex(msg.OutTxChain, msg.OutTxTssNonce), ctx.BlockHeight()))
-	k.SetCctxAndNonceToCctxAndInTxHashToCctx(ctx, cctx)
+	k.SaveSucessfullOutBound(ctx, &cctx, msg.ValueReceived.String())
 	return &types.MsgVoteOnObservedOutboundTxResponse{}, nil
-}
-
-func percentOf(n *big.Int, percent int64) *big.Int {
-	n = n.Mul(n, big.NewInt(percent))
-	n = n.Div(n, big.NewInt(100))
-	return n
-}
-
-// FundGasStabilityPoolFromRemainingFees funds the gas stability pool with the remaining fees of an outbound tx
-func (k Keeper) FundGasStabilityPoolFromRemainingFees(ctx sdk.Context, outboundTxParams types.OutboundTxParams, chainID int64) error {
-	gasUsed := outboundTxParams.OutboundTxGasUsed
-	gasLimit := outboundTxParams.OutboundTxEffectiveGasLimit
-	gasPrice := math.NewUintFromBigInt(outboundTxParams.OutboundTxEffectiveGasPrice.BigInt())
-
-	if gasLimit == gasUsed {
-		return nil
-	}
-
-	// We skip gas stability pool funding if one of the params is zero
-	if gasLimit > 0 && gasUsed > 0 && !gasPrice.IsZero() {
-		if gasLimit > gasUsed {
-			remainingGas := gasLimit - gasUsed
-			remainingFees := math.NewUint(remainingGas).Mul(gasPrice).BigInt()
-
-			// We fund the stability pool with a portion of the remaining fees
-			remainingFees = percentOf(remainingFees, RemainingFeesToStabilityPoolPercent)
-
-			// Fund the gas stability pool
-			if err := k.fungibleKeeper.FundGasStabilityPool(ctx, chainID, remainingFees); err != nil {
-				return err
-			}
-		} else {
-			return fmt.Errorf("VoteOnObservedOutboundTx: The gas limit %d is less than the gas used %d", gasLimit, gasUsed)
-		}
-	}
-	return nil
 }
